+++ python_modules/dagster/dagster_tests/core_tests/definitions_tests/test_decorators.py
@@ -178,6 +178,7 @@ def test_solid_with_input():
 
     pipeline = PipelineDefinition(
         solid_defs=[define_stub_solid("test_value", {"foo": "bar"}), hello_world],
+        name="test",
         dependencies={"hello_world": {"foo_to_foo": DependencyDefinition("test_value")}},
     )
 
@@ -257,12 +258,12 @@ def test_wrong_argument_to_pipeline():
     with pytest.raises(
         DagsterInvalidDefinitionError, match="You have passed a lambda or function non_solid_func"
     ):
-        PipelineDefinition(solid_defs=[non_solid_func])
+        PipelineDefinition(solid_defs=[non_solid_func], name="test")
 
     with pytest.raises(
         DagsterInvalidDefinitionError, match="You have passed a lambda or function <lambda>"
     ):
-        PipelineDefinition(solid_defs=[lambda x: x])
+        PipelineDefinition(solid_defs=[lambda x: x], name="test")
 
 
 def test_descriptions():