+++ python_modules/dagster/dagster_tests/core_tests/execution_tests/test_timing.py
@@ -14,7 +14,7 @@ def test_event_timing_before_yield():
         time.sleep(0.01)
         yield Output(None)
 
-    pipeline_def = PipelineDefinition(solid_defs=[before_yield_solid])
+    pipeline_def = PipelineDefinition(solid_defs=[before_yield_solid], name="test")
     pipeline_result = execute_pipeline(pipeline_def)
     success_event = pipeline_result.result_for_solid("before_yield_solid").get_step_success_event()
     assert success_event.event_specific_data.duration_ms >= 10.0
@@ -29,7 +29,7 @@ def test_event_timing_after_yield():
         yield Output(None)
         time.sleep(0.01)
 
-    pipeline_def = PipelineDefinition(solid_defs=[after_yield_solid])
+    pipeline_def = PipelineDefinition(solid_defs=[after_yield_solid], name="test")
     pipeline_result = execute_pipeline(pipeline_def)
     success_event = pipeline_result.result_for_solid("after_yield_solid").get_step_success_event()
     assert success_event.event_specific_data.duration_ms >= 10.0
@@ -44,7 +44,7 @@ def test_event_timing_direct_return():
         time.sleep(0.01)
         return None
 
-    pipeline_def = PipelineDefinition(solid_defs=[direct_return_solid])
+    pipeline_def = PipelineDefinition(solid_defs=[direct_return_solid], name="test")
     pipeline_result = execute_pipeline(pipeline_def)
     success_event = pipeline_result.result_for_solid("direct_return_solid").get_step_success_event()
     assert success_event.event_specific_data.duration_ms >= 10.0