+++ python_modules/dagster/dagster_tests/core_tests/test_input_injection.py
@@ -43,6 +43,7 @@ def test_string_from_aliased_inputs():
 
     pipeline = PipelineDefinition(
         solid_defs=[str_as_input],
+        name="test",
         dependencies={SolidInvocation("str_as_input", alias="aliased"): {}},
     )