+++ python_modules/dagster/dagster_tests/core_tests/test_mode_definitions.py
@@ -1,5 +1,4 @@
 import logging
-import re
 
 import pytest
 from dagster import (